+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+- Add examples
+
 ## 1.2.0 - 2026-01-30
 ### Added
 - Add python and ruby tests